feat: rename SrcInfo imports
