+++ src/abi/AbiFunction.ts
@@ -1,7 +1,8 @@
-import { AstExpression, SrcInfo } from "../grammar/ast";
+import { AstExpression } from "../grammar/ast";
 import { CompilerContext } from "../context";
 import { WriterContext } from "../generator/Writer";
 import { TypeRef } from "../types/types";
+import { SrcInfo } from "../grammar/prev";
 
 export type AbiFunction = {
     name: string;
