+++ src/errors.ts
@@ -1,8 +1,8 @@
 import { MatchResult } from "ohm-js";
 import path from "path";
 import { cwd } from "process";
-import { AstFuncId, AstId, AstTypeId, SrcInfo } from "./grammar/ast";
-import { ItemOrigin } from "./grammar/prev";
+import { AstFuncId, AstId, AstTypeId } from "./grammar/ast";
+import { ItemOrigin, SrcInfo } from "./grammar/prev";
 
 export class TactError extends Error {
     readonly loc?: SrcInfo;
