+++ src/grammar/prev/checkConstAttributes.ts
@@ -1,5 +1,6 @@
-import { AstConstantAttribute, SrcInfo } from "../ast";
+import { AstConstantAttribute } from "../ast";
 import { throwSyntaxError } from "../../errors";
+import { SrcInfo } from "./src-info";
 
 export function checkConstAttributes(
     isAbstract: boolean,
