+++ src/types/resolveStatements.ts
@@ -1,7 +1,6 @@
 import { CompilerContext } from "../context";
 import {
     AstCondition,
-    SrcInfo,
     AstStatement,
     tryExtractPath,
     AstId,
@@ -31,6 +30,7 @@ import { FunctionDescription, printTypeRef, TypeRef } from "./types";
 import { evalConstantExpression } from "../constEval";
 import { ensureInt } from "../interpreter";
 import { crc16 } from "../utils/crc16";
+import { SrcInfo } from "../grammar/prev";
 
 export type StatementContext = {
     root: SrcInfo;
